requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
certifi==2021.5.30
    # via requests
charset-normalizer==2.0.4
    # via requests
factory-boy==3.3.0
    # via -r requirements.in
faker==20.0.0
    # via factory-boy
freezegun==1.2.2
    # via -r requirements.in
idna==3.2
    # via requests
python-dateutil==2.8.2
    # via
    #   faker
    #   freezegun
requests==2.26.0
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
urllib3==1.26.6
    # via requests
